Move util function to generic new file
