+++ libvirt/utils_test.go
@@ -3,30 +3,12 @@ package libvirt
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"log"
 	"net"
 	"os"
 	"testing"
 	"time"
-
-	"github.com/hashicorp/terraform/terraform"
 )
-
-// getResourceFromTerraformState is helper function for getting a resource by name
-// from terraform states produced during testacc
-func getResourceFromTerraformState(resourceName string, state *terraform.State) (*terraform.ResourceState, error) {
-	rs, ok := state.RootModule().Resources[resourceName]
-	if !ok {
-		return nil, fmt.Errorf("Not found: %s", resourceName)
-	}
-
-	if rs.Primary.ID == "" {
-		return nil, fmt.Errorf("No libvirt resource key ID is set")
-	}
-	return rs, nil
-
-}
 
 func TestDiskLetterForIndex(t *testing.T) {
